tslint.json -> tslint.js
The hardcoded rulesDirectory path to tslint-eslint-rules doesn't
work without shrinkwrap because of dependency flattening.

If for some reason we don't want to do this and don't want to
shrinkwrap, we could extend tslint-eslint-rules, but then require
that the root project add `node_modules/tslint-eslint-rules/dist/rules`
to rulesDirectory in its tslint config.

Additional info:
https://github.com/buzinas/tslint-eslint-rules/issues/115